* origin/wlan-cmn.driver.lnx.1.0-dev:
  qcacmn: Add bus_id memeber to qdf_device
  Release 5.1.0.22K
  qcacmn: Initialize ce ring base address high registers for adrastea
  qcacmn: Fix Q_TARGET_ACCESS in hif_fastpath_resume
  Release 5.1.0.22J
  qcacmn: Compilation fix in DP layer caused due to transport layer
  qcacmn: Compilation fixes for AP135 platform
  Release 5.1.0.22I
  qcacmn: Fix compilation with WMI_INTERFACE_EVENT_LOGGING disabled
  Release 5.1.0.22H
  qcacmn: Add unit test framework to validate WLAN suspend/resume features
  Release 5.1.0.22G
  qcacmn: Remove unnecessary locking during WMI detach
  Release 5.1.0.22F
  qcacmn: CL 1597848 - update fw common interface files
  Release 5.1.0.22E
  qcacmn: Correct len size of oem data request

+++ wmi/src/wmi_unified_api.c
@@ -2264,7 +2264,7 @@ QDF_STATUS wmi_unified_csa_offload_enable(void *wmi_hdl, uint8_t vdev_id)
  * Return: QDF_STATUS_SUCCESS on success and QDF_STATUS_E_FAILURE for failure
  */
 QDF_STATUS wmi_unified_start_oem_data_cmd(void *wmi_hdl,
-			  uint8_t data_len,
+			  uint32_t data_len,
 			  uint8_t *data)
 {
 	wmi_unified_t wmi_handle = (wmi_unified_t) wmi_hdl;
